+++ src/tacticode/engine/BattleEngine.cpp
@@ -125,7 +125,7 @@ namespace tacticode
 		}
 		std::shared_ptr<Character> BattleEngine::getCharacter(int32_t characterId)
 		{
-			for (int i = 0; i < m_characters.size(); ++i)
+			for (size_t i = 0; i < m_characters.size(); ++i)
 			{
 				if (m_characters[i]->getId() == characterId)
 					return (m_characters[i]);
